Add files via upload
Title: Random Password Generator

Description:
This Python script is a versatile random password generator that creates strong and secure passwords tailored to your specifications. You can customize the minimum length of the password and choose whether to include numbers and special characters. The generated passwords are robust and suitable for protecting sensitive information in various contexts, from online accounts to confidential documents. With its easy-to-use interface and reliable functionality, this tool ensures that your digital assets remain safe and secure.